+++ container/stack.h
@@ -16,8 +16,8 @@
 typedef char STACKTYPE;
 typedef struct {
 	STACKTYPE* top;
-	unsigned int size;
-	unsigned int n;
+	uint16_t size;
+	uint16_t n;
 } STACK;
 /* Exported constants --------------------------------------------------------*/
 #define STACK_USEAUTORESIZING 0
@@ -26,12 +26,12 @@ typedef struct {
 #define STACK_STEP 16
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions ------------------------------------------------------- */
-STACK* stack_create(int size);
+STACK* stack_create(uint16_t size);
 void stack_push(STACK* stack, STACKTYPE el);
 STACKTYPE stack_pop(STACK* stack);
 BOOL stack_empty(STACK* stack);
 void stack_destroy(STACK* stack);
-void stack_resize(STACK* stack, unsigned n);
+void stack_resize(STACK* stack, uint16_t n);
 unsigned stack_num(STACK* stack);
 
 #endif /* _H_STACK */
