+++ dsp/dft.h
@@ -12,6 +12,7 @@
 #define _H_DFT
 
 /* Includes ------------------------------------------------------------------*/
+#include <stdint.h>
 #include "../common/util.h"
 #include "../common/complex.h"
 /* Exported types ------------------------------------------------------------*/
@@ -19,13 +20,13 @@
 #define DEL_ROUND_ERR 0
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions ------------------------------------------------------- */
-void dft(UTILTYPE *aSamples, COMPLEX *spectrum, unsigned int N);
-void dft_both(UTILTYPE *aSamples, UTILTYPE *aMagnitude, UTILTYPE *aPhase, unsigned int N);
-void dft_mag(UTILTYPE *aSamples, UTILTYPE *aMagnitude, unsigned int N);
-void dft_phase(UTILTYPE *aSamples, UTILTYPE *aPhase, unsigned int N);
-COMPLEX dft_k(UTILTYPE *aSamples, unsigned k, unsigned N);
-void dft_both_k(UTILTYPE *aSamples, UTILTYPE* magnitude, UTILTYPE* phase, unsigned k, unsigned N);
-UTILTYPE dft_mag_k(UTILTYPE *aSamples, unsigned k, unsigned N);
-UTILTYPE dft_phase_k(UTILTYPE *aSamples, unsigned k, unsigned N);
+void dft(UTILTYPE *aSamples, COMPLEX *spectrum, uint16_t N);
+void dft_both(UTILTYPE *aSamples, UTILTYPE *aMagnitude, UTILTYPE *aPhase, uint16_t N);
+void dft_mag(UTILTYPE *aSamples, UTILTYPE *aMagnitude, uint16_t N);
+void dft_phase(UTILTYPE *aSamples, UTILTYPE *aPhase, uint16_t N);
+COMPLEX dft_k(UTILTYPE *aSamples, uint16_t k, uint16_t N);
+void dft_both_k(UTILTYPE *aSamples, UTILTYPE* magnitude, UTILTYPE* phase, uint16_t k, uint16_t N);
+UTILTYPE dft_mag_k(UTILTYPE *aSamples, uint16_t k, uint16_t N);
+UTILTYPE dft_phase_k(UTILTYPE *aSamples, uint16_t k, uint16_t N);
 
 #endif /* _H_DFT */
