+++ dsp/fir.c
@@ -6,15 +6,15 @@
  */
 /* Includes ------------------------------------------------------------------*/
 #include "fir.h"
-#include <malloc.h>
+#include <stdlib.h>
 /* Private typedef -----------------------------------------------------------*/
 /* Private define ------------------------------------------------------------*/
 /* Private macro -------------------------------------------------------------*/
 /* Private variables ---------------------------------------------------------*/
 volatile FIRTYPE *fir_coeff;
-volatile unsigned *fir_coeff_n;
+volatile uint16_t *fir_coeff_n;
 volatile FIRTYPE *fir_samples_history;
-volatile unsigned *fir_oldest_sample;
+volatile uint16_t *fir_oldest_sample;
 /* Private function prototypes -----------------------------------------------*/
 /* Private functions ---------------------------------------------------------*/
 
@@ -26,12 +26,12 @@ volatile unsigned *fir_oldest_sample;
 * Output         : None.
 * Return         : None.
 *******************************************************************************/
-void fir_init(FIRTYPE *coeff, unsigned n) {
-	unsigned i;
+void fir_init(FIRTYPE *coeff, uint16_t n) {
+	uint16_t i;
 	fir_coeff = (FIRTYPE*) malloc(n*sizeof(FIRTYPE));
-	fir_coeff_n = (unsigned*) malloc(sizeof(unsigned));
+	fir_coeff_n = (uint16_t*) malloc(sizeof(uint16_t));
 	fir_samples_history = (FIRTYPE*) calloc(n, sizeof(FIRTYPE));
-	fir_oldest_sample = (unsigned*) malloc(sizeof(unsigned));
+	fir_oldest_sample = (uint16_t*) malloc(sizeof(uint16_t));
 	*fir_oldest_sample = 0;
 	*fir_coeff_n = n;
 	for(i = 0; i < n; i++) {
@@ -61,7 +61,7 @@ void fir_free() {
 * Return         : None.
 *******************************************************************************/
 void fit_clear_input() {
-	unsigned i, N;
+	uint16_t i, N;
 	N = *fir_coeff_n;
 	for(i = 0; i < N; i++) {
 		*(fir_samples_history+i) = 0;
@@ -76,7 +76,7 @@ void fit_clear_input() {
 * Return         : Output sample.
 *******************************************************************************/
 FIRTYPE fir(FIRTYPE sample) {
-	unsigned i, N;
+	uint16_t i, N;
 	*(fir_samples_history+(*fir_oldest_sample)) = sample;
 	N = *fir_coeff_n;
 	sample = 0;
